backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
langchain-community==0.0.XX
langchain-core==0.0.XX
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.22.0
chromadb==0.4.0
sentence-transformers==2.2.2
google-generativeai==0.6.0